the cake is a lie hack

+++ src/main/groovy/com/rodhilton/rectanglevisibility/domain/Rectangle.groovy
@@ -3,7 +3,7 @@ package com.rodhilton.rectanglevisibility.domain
 class Rectangle implements Serializable {
     static final long serialVersionUID = 50L;
 
-    int east, north, west, south
+    double east, north, west, south
 
     def String toString() {
         "(${east}, ${north}, ${west}, ${south})"
